# oas3_v2.yaml
openapi: 3.0.0
info:
  version: '2.0'
  title: "TrentoInBici OpenAPI 3.0"
  description: API for moving and parking the bicycle in Trento. 
  license:
    name: MIT
servers:
  - url: https://trentoinbici.onrender.com/api/v2
    description: trentoinbici.onrender.com
paths:
...
  /adminDatabase:
      get:
        description: >-
          Return all reported racks
        summary: Return all reported racks
        responses:
          '200':
            description: All reported racks
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Position received successfully
                    body:
                      type: array
                      items:
                        $ref: '#/components/schemas/Rack'
        parameters:
          - in: header
            name: x-access-token
            schema:
              type: string
            required: true
            description: Access token for authentication
  /adminDatabase/add:
    post:
      description: >-
        Receives the rack to add to the database
      summary: Receives the rack to add to the database
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: object
                  properties:
                    id:
                    type: string
                    example: "1"
                    latitude:
                      type: number
                      format: double
                      description: Latitude of position.
                      example: 46.7128
                    longitude:
                      type: number
                      format: double
                      description: Longitude of position.
                      example: 11.1211
                  required:
                    - latitude
                    - longitude
      parameters:
        - in: header
          name: x-access-token
          schema:
            type: string
          required: true
          description: Access token for authentication
      responses:
        '200':
          description: Successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Racks added successfully
  /adminDatabase/noAdd:
    post:
      description: >-
        Receives the rack to delete from the database
      summary: Receives the rack to delete from the database
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: object
                  properties:
                    id:
                    type: string
                    example: "1"
                    latitude:
                      type: number
                      format: double
                      description: Latitude of position.
                      example: 46.7128
                    longitude:
                      type: number
                      format: double
                      description: Longitude of position.
                      example: 11.1211
                  required:
                    - latitude
                    - longitude
      parameters:
        - in: header
          name: x-access-token
          schema:
            type: string
          required: true
          description: Access token for authentication
      responses:
        '200':
          description: Successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Racks deleted successfully
...
components:
  schemas:
    Rack:
      type: object
      properties:
        id:
          type: string
          description: racks Id.
          example: "1"
        latitude:
          type: number
          format: double
          description: Latitude
          example: 46.7128
        longitude:
          type: number
          format: double
          description: Longitude 
          example: 11.1211
        distance:
          type: number
          format: double
          description: distance position-rack
        travelTime:
          type: number
          format: double
          description: travel time position-rack
...